Install Rust
I'm using the rust-lang binary packages by default so that it's possible
to build a Rust machine without having to have compilers etc installed
on production machines which can be a security risk in some
organisations.

Version and architecture is configured from the recipe attributes so you
can install a specific version or the nightlies.