Support CentOS
In order to support the latest CentOS we need to set the
$LD_LIBRARY_PATH to pull in libraries from `usr/local/lib`. This path
*seems* to already be set on Ubuntu although I'm not sure how or where.

This commit set's it up explicitly so that CentOS can find the libs and
runs `ldconfig` to make sure that the system is set up properly.

This approach works on both systems, and my preference is to be explicit
about what paths are loaded rather than implicit.